+++ src/constants.ts
@@ -29,12 +29,26 @@ export const GLOB_JSON = '**/*.json' as const;
 export const GLOB_JSONC = '**/*.jsonc' as const;
 export const GLOB_JSON5 = '**/*.json5' as const;
 
-export const GLOB_YAML = '**/*.y?(a)ml' as const;
+export const GLOB_YAML_EXTENSION = 'y?(a)ml' as const;
+export const GLOB_YAML = `**/*.${GLOB_YAML_EXTENSION}` as const;
 
 export const GLOB_TOML = '**/*.toml' as const;
 
 export const GLOB_PACKAGE_JSON = '**/package.json' as const;
 
+export const GLOB_SUPPORTED_EXTENSIONS = [
+  GLOB_JS_TS_EXTENSION,
+  'vue',
+  'json',
+  'jsonc',
+  'json5',
+  GLOB_YAML_EXTENSION,
+  'toml',
+].join(',');
+export const GLOB_MARKDOWN = '**/*.md' as const;
+export const GLOB_MARKDOWN_SUPPORTED_CODE_BLOCKS =
+  `${GLOB_MARKDOWN}/**/*.{${GLOB_SUPPORTED_EXTENSIONS}}` as const;
+export const GLOB_MARKDOWN_ALL_CODE_BLOCKS = `${GLOB_MARKDOWN}/**/*.*` as const;
 
 /* Misc */