+++ src/types/utils.ts
@@ -19,4 +19,7 @@ export type ConstantKeys<T> = {
 
 export type PrettifyShallow<T> = {
   [K in keyof T]: T[K];
-};
+} & {};
+export type Prettify<T> = {
+  [K in keyof T]: Prettify<T[K]>;
+} & {};