+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Unreleased
 
-- Fixed an issue in `vue` config where Nuxt's `app.vue` and `error.vue` files in a custom project directory were not recognized.
+- New config: `markdown` via [`@eslint/markdown`](https://www.npmjs.com/package/@eslint/markdown), enabled by default.
 - Added a boolean option `overrideIgnores` to completely override global `ignores` provided by our config.
 - Added a `node` config option to configure [`prefer-global`](https://github.com/eslint-community/eslint-plugin-n/tree/master/docs/rules/prefer-global) rule for each feature.
+- Fixed an issue in `vue` config where Nuxt's `app.vue` and `error.vue` files in a custom project directory were not recognized.
 - Update dependencies.
     
 ## 0.1.6